backend/requirements.txt
fastapi
uvicorn
nemo_toolkit[asr]
python-multipart
